refactor: decouple session hq from legacy SessionView
