+++ src/renderer/src/components/sessions/index.ts
@@ -10,5 +10,5 @@ export { StreamingCursor } from './StreamingCursor'
 export { CodeBlock } from './CodeBlock'
 export { MarkdownRenderer } from './MarkdownRenderer'
 export { ModeToggle } from './ModeToggle'
-export type { OpenCodeMessage, SessionViewState, StreamingPart } from './SessionView'
+export type { OpenCodeMessage, SessionViewState, StreamingPart } from '@/lib/session-types'
 export type { ToolStatus, ToolUseInfo } from './ToolCard'
